# fasofree-backend/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Obtenir le profil de l’utilisateur connecté",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users": {
      "get": {
        "operationId": "UsersController_getAllUsers",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lister les utilisateurs (super administrateur)",
        "tags": [
          "Users"
        ]
      },
      "post": {
        "operationId": "UsersController_createUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Créer un utilisateur (SUPER_ADMIN) — seul moyen de créer un ADMIN/SUPPORT",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/status": {
      "patch": {
        "operationId": "UsersController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Bannir ou réactiver un compte (SUPER_ADMIN)",
        "tags": [
          "Users"
        ]
      }
    },
    "/api/v1/users/{id}/role": {
      "patch": {
        "operationId": "UsersController_updateRole",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Changer le rôle d’un utilisateur (SUPER_ADMIN)",
        "tags": [
          "Users"
        ]
      }
    },
...
    "/api/v1/subscriptions/plans": {
      "post": {
        "operationId": "SubscriptionsController_createPlan",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[Admin] Créer un forfait",
        "tags": [
          "Subscriptions"
        ]
      },
      "get": {
        "operationId": "SubscriptionsController_getPlans",
        "parameters": [
          {
            "name": "subjectType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Lister les forfaits disponibles (catalogue)",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/plans/{code}": {
      "patch": {
        "operationId": "SubscriptionsController_updatePlan",
        "parameters": [
          {
            "name": "code",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[Admin] Modifier un forfait (prix, durée, commission, avantages)",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions": {
      "get": {
        "operationId": "SubscriptionsController_listSubscriptions",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[Admin] Lister tous les abonnements",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/assign": {
      "post": {
        "operationId": "SubscriptionsController_assignSubscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[Admin] Assigner ou renouveler un forfait à un commerce/client",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/renew": {
      "post": {
        "operationId": "SubscriptionsController_renewSubscription",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenewSubscriptionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "[Admin] Renouveler l’abonnement actif d’un sujet",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/me": {
      "get": {
        "operationId": "SubscriptionsController_getMyStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Mon statut premium (FasoFree Pass)",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/subscribe": {
      "post": {
        "operationId": "SubscriptionsController_subscribe",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "S’abonner au FasoFree Pass (débit portefeuille)",
        "tags": [
          "Subscriptions"
        ]
      }
    },
    "/api/v1/subscriptions/merchant/subscribe": {
      "post": {
        "operationId": "SubscriptionsController_subscribeMerchant",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MerchantSubscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "JWT-auth": []
          }
        ],
        "summary": "Abonner son commerce au plan Pro (débit portefeuille marchand)",
        "tags": [
          "Subscriptions"
        ]
      }
    },
...
  },
  "info": {
    "title": "FasoFree API",
    "description": "Spécification OpenAPI officielle du backend FasoFree - Marketplace & Livraison à Ouagadougou",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "Serveur courant"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "Authorization",
        "description": "Saisissez votre token JWT",
        "in": "header"
      }
    },
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Awa Diallo"
          },
          "email": {
            "type": "string",
            "example": "awa@fasofree.bf",
            "format": "email"
          },
          "phone": {
            "type": "string",
            "example": "+22670000000"
          },
          "password": {
            "type": "string",
            "example": "MotDePasseFort123!",
            "format": "password"
          },
          "role": {
            "type": "string",
            "enum": [
              "client",
              "courier",
              "driver",
              "business_admin",
              "support",
              "admin",
              "super_admin"
            ],
            "example": "support"
          }
        },
        "required": [
          "fullName",
          "email",
          "phone",
          "password"
        ]
      },
      "UpdateUserStatusDto": {
        "type": "object",
        "properties": {
          "isActive": {
            "type": "boolean",
            "example": false,
            "description": "true = actif, false = banni"
          }
        },
        "required": [
          "isActive"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "enum": [
              "client",
              "courier",
              "driver",
              "business_admin",
              "support",
              "admin",
              "super_admin"
            ],
            "example": "admin"
          }
        },
        "required": [
          "role"
        ]
      },
...
      "CreatePlanDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VIP",
            "description": "Code unique du forfait (majuscules)"
          },
          "name": {
            "type": "string",
            "example": "FasoFree Pass VIP"
          },
          "subjectType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "MERCHANT"
            ],
            "example": "CUSTOMER"
          },
          "priceFcfa": {
            "type": "number",
            "example": 2500
          },
          "durationDays": {
            "type": "number",
            "example": 30
          },
          "commissionRate": {
            "type": "object",
            "example": 0.015,
            "description": "Taux de commission préférentiel (ex: 0.015 = 1.5%), null = standard"
          },
          "freeServiceFee": {
            "type": "boolean",
            "default": false
          },
          "freeDelivery": {
            "type": "boolean",
            "default": false
          },
          "freeDeliveryMinSubtotal": {
            "type": "number",
            "default": 0,
            "description": "Seuil de panier pour livraison gratuite"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "code",
          "name",
          "subjectType",
          "priceFcfa"
        ]
      },
      "UpdatePlanDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "VIP",
            "description": "Code unique du forfait (majuscules)"
          },
          "name": {
            "type": "string",
            "example": "FasoFree Pass VIP"
          },
          "subjectType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "MERCHANT"
            ],
            "example": "CUSTOMER"
          },
          "priceFcfa": {
            "type": "number",
            "example": 2500
          },
          "durationDays": {
            "type": "number",
            "example": 30
          },
          "commissionRate": {
            "type": "object",
            "example": 0.015,
            "description": "Taux de commission préférentiel (ex: 0.015 = 1.5%), null = standard"
          },
          "freeServiceFee": {
            "type": "boolean",
            "default": false
          },
          "freeDelivery": {
            "type": "boolean",
            "default": false
          },
          "freeDeliveryMinSubtotal": {
            "type": "number",
            "default": 0,
            "description": "Seuil de panier pour livraison gratuite"
          },
          "description": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "AssignSubscriptionDto": {
        "type": "object",
        "properties": {
          "subjectType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "MERCHANT"
            ],
            "example": "MERCHANT"
          },
          "subjectId": {
            "type": "string",
            "description": "ID du Business (commerçant) ou du User (client)"
          },
          "planCode": {
            "type": "string",
            "example": "PRO"
          },
          "durationDays": {
            "type": "number",
            "description": "Durée en jours (défaut : durée du forfait)"
          },
          "autoRenew": {
            "type": "boolean",
            "default": true
          },
          "renew": {
            "type": "boolean",
            "default": false,
            "description": "true = prolonger l’abonnement actif"
          },
          "debitWallet": {
            "type": "boolean",
            "default": false,
            "description": "true = débiter le portefeuille marchand du prix du forfait (mode déductible)"
          }
        },
        "required": [
          "subjectType",
          "subjectId",
          "planCode"
        ]
      },
      "RenewSubscriptionDto": {
        "type": "object",
        "properties": {
          "subjectType": {
            "type": "string",
            "enum": [
              "CUSTOMER",
              "MERCHANT"
            ]
          },
          "subjectId": {
            "type": "string"
          },
          "durationDays": {
            "type": "number"
          }
        },
        "required": [
          "subjectType",
          "subjectId"
        ]
      },
      "SubscribeDto": {
        "type": "object",
        "properties": {
          "planCode": {
            "type": "string",
            "default": "VIP",
            "description": "Code du forfait client (ex: VIP)"
          },
          "autoRenew": {
            "type": "boolean",
            "default": true
          }
        }
      },
      "MerchantSubscribeDto": {
        "type": "object",
        "properties": {
          "businessId": {
            "type": "string",
            "description": "ID du commerce à abonner"
          },
          "planCode": {
            "type": "string",
            "default": "PRO",
            "description": "Code du forfait marchand (ex: PRO)"
          },
          "autoRenew": {
            "type": "boolean",
            "default": true
          }
        },
        "required": [
          "businessId"
        ]
      },
...
    }
  }
}